# Faster `min`, `max`, `minmax`