@stevengj 2021-12-17
> I think that one `isnan` check in `min` / `max` is redundant, as `y < x` will be `false` if `x` is `NaN`. 

The `signbit` check can still be `true`, though.  With the implementation in this PR, I get the incorrect result
```jl
julia> newmax(-NaN, 1.0)
1.0
```