@mcabbott 2021-12-17
Oh that's awful, there's `-NaN`. But thanks, that's what I was missing.