@oscardssmith 2021-12-17
`-NaN` is a `NaN`, but there are `2^52` different `NaN` values.