@mcabbott 2021-12-17
Could `signbit` easily check one more bit, so as to give them a consistent answer?